backend/pyproject.toml
[project]
name = "arxiviz-backend"
version = "0.1.0"
description = "ArXiviz Backend - Team 2 AI Agent Pipeline for Manim Visualization Generation"
requires-python = ">=3.11"
dependencies = [
    # API framework
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Core AI + validation
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.21.1,<0.22.0",

    # Async / transport
    "httpx>=0.25.0",
    "asyncio-throttle>=1.0.0",

    # Database / jobs
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",

    # Ingestion (Team 1)
    "arxiv>=2.0.0",
    "pymupdf4llm>=0.0.5",
    "pymupdf>=1.23.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # Rendering / visualization
    "manim>=0.18.0",
    "modal>=0.60.0",
    "manim-voiceover[elevenlabs]>=0.3.0",

    # LLM
    "openai>=1.12.0",

    # Dedalus SDK + MCP Gateway (Context7 live docs)
    "dedalus-labs>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "prompts", "examples"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.workspace]
members = [
    "manimations",
]
